Pass the hart ID on start_kernel instead
Instead of storing the hard ID as given by the bootloader into a C
variable, pass it directly into the 'start_kernel' function since we
don't need to tamper the value as originally laid out on the 'a0'
register.

+++ include/fbos/printk.h
@@ -13,7 +13,7 @@
 extern void die(const char *const message);
 
 // Print the given number as a single digit.
-extern void print_digit(uint32_t digit);
+extern void print_digit(uint64_t digit);
 
 // Print the given message.
 extern void printk(const char *const message);
